Build out tests for bounding box  - Includes initial reorg of code
