+++ packages/fell_finder/Cargo.toml
@@ -14,6 +14,7 @@ indicatif = "0.17.11"
 axum = "0.8.1"
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.140"
+approx = "0.5.1"
 
 [profile.release]
 opt-level = 3
