--- packages/fell_finder/src/config.rs
@@ -1,1 +1,0 @@
-pub mod route;
