+++ packages/fell_finder/src/lib.rs
@@ -1,3 +1,3 @@
-pub mod config;
+pub mod common;
 pub mod loading;
 pub mod routing;
