+++ packages/fell_finder/src/loading.rs
@@ -1,3 +1,2 @@
 pub mod petgraph;
 pub mod postgres;
-pub mod structs;
