+++ packages/fell_finder/src/loading/postgres.rs
@@ -1,7 +1,69 @@
-use crate::config::route::RouteConfig;
-use crate::loading::structs::{EdgeRow, NodeRow};
+use crate::common::config::RouteConfig;
+use crate::common::graph_data::{EdgeData, NodeData};
 use sqlx;
 use sqlx::PgPool;
+
+/// Container for the raw output of the nodes SQL query
+#[derive(sqlx::FromRow, Debug, Clone, Copy)]
+pub struct NodeRow {
+    pub id: i64,
+    lat: f64,
+    lon: f64,
+    elevation: f64,
+}
+
+impl NodeRow {
+    /// Unpack the raw node data into a format which can be loaded into the
+    /// graph
+    pub fn prepare(self) -> NodeData {
+        NodeData {
+            id: self.id,
+            lat: self.lat,
+            lon: self.lon,
+            elevation: self.elevation,
+            is_start: false, // default, will be overwritten``
+            dist_to_start: None, // default, will be overwritten
+        }
+    }
+}
+
+/// Container for the raw output of the edges SQL query
+#[derive(sqlx::FromRow, Debug)]
+pub struct EdgeRow {
+    pub src: i64,
+    pub dst: i64,
+    highway: String,
+    surface: String,
+    elevation_gain: f64,
+    elevation_loss: f64,
+    distance: f64,
+    lats: Vec<f64>,
+    lons: Vec<f64>,
+    eles: Vec<f64>,
+    dists: Vec<f64>,
+}
+
+impl EdgeRow {
+    /// Unpack the raw edge data into a format which can be loaded into the
+    /// graph
+    pub fn prepare(self) -> (i64, i64, EdgeData) {
+        let data = EdgeData {
+            src: self.src,
+            dst: self.dst,
+            highway: self.highway,
+            surface: self.surface,
+            elevation_gain: self.elevation_gain,
+            elevation_loss: self.elevation_loss,
+            distance: self.distance,
+            lats: self.lats,
+            lons: self.lons,
+            eles: self.eles,
+            dists: self.dists,
+        };
+
+        (self.src, self.dst, data)
+    }
+}
 
 /// Generate a SQL query to read in the nodes for a route based on the
 /// provided user configuration
@@ -11,20 +73,15 @@ pub fn generate_nodes_query(config: &RouteConfig) -> String {
     let ptn_list = bbox.get_partition_list();
     let ptn_str = ptn_list.join(", ");
 
-    let min_lat = &bbox.sw.y().to_string();
-    let min_lon = &bbox.sw.x().to_string();
-    let max_lat = &bbox.ne.y().to_string();
-    let max_lon = &bbox.ne.x().to_string();
-
     // This brings the query into the compiled code
     let nodes_base = include_str!("get_nodes.sql");
 
     let nodes_query = nodes_base
         .replace("< ptn_str >", &ptn_str)
-        .replace("< min_lat >", &min_lat)
-        .replace("< min_lon >", &min_lon)
-        .replace("< max_lat >", &max_lat)
-        .replace("< max_lon >", &max_lon);
+        .replace("< min_lat >", &bbox.min_lat.to_string())
+        .replace("< min_lon >", &bbox.min_lon.to_string())
+        .replace("< max_lat >", &bbox.max_lat.to_string())
+        .replace("< max_lon >", &bbox.max_lon.to_string());
 
     nodes_query
 }
@@ -35,11 +92,9 @@ pub fn generate_nodes_query(config: &RouteConfig) -> String {
 pub fn generate_edges_query(config: &RouteConfig) -> String {
     let bbox = config.get_bounding_box();
 
-    // TODO: Set BBox class to call this once at instantiation
     let ptn_list = bbox.get_partition_list();
     let ptn_str = ptn_list.join(", ");
 
-    let coords = bbox.get_min_max_coords();
     let surface_str = config.get_surface_str();
     let highway_str = config.get_highway_str();
 
@@ -49,10 +104,10 @@ pub fn generate_edges_query(config: &RouteConfig) -> String {
         .replace("< ptn_str >", &ptn_str)
         .replace("< highway_str >", &highway_str)
         .replace("< surface_str >", &surface_str)
-        .replace("< min_lat >", &coords.min_lat.to_string())
-        .replace("< min_lon >", &coords.min_lon.to_string())
-        .replace("< max_lat >", &coords.max_lat.to_string())
-        .replace("< max_lon >", &coords.max_lon.to_string());
+        .replace("< min_lat >", &bbox.min_lat.to_string())
+        .replace("< min_lon >", &bbox.min_lon.to_string())
+        .replace("< max_lat >", &bbox.max_lat.to_string())
+        .replace("< max_lon >", &bbox.max_lon.to_string());
 
     edges_query
 }
